rename(cai): rename module from 'cai' to 'c2pa'
