+++ setup.py
@@ -9,8 +9,8 @@ with open("README.md", "r") as f:
 
 
 setup(
-    name='cai',
-    version='v1.3.2',
+    name='c2pa',
+    version='v2.1.0',
     description='Content Authenticity Initiative Implementation.',
     long_description=long_description,
     long_description_content_type="text/markdown",
@@ -35,14 +35,13 @@ setup(
         'pymultihash>=0.8.2',
         'py3exiv2>=0.9.3',
         'endesive>=2.0.2',
-        #'tox',
         'cbor>=1.0.0',
     ],
     python_requires='>=3',
     entry_points={
         'console_scripts': [
-            'cai_tool=cai.cai_tool:main',
-            'starling_cai=cai.starling:main',
+            'cai_tool=c2pa.cai_tool:main',
+            'c2pa=c2pa.starling:main',
         ]
     },
     test_suite='tests'
