+++ utils/starling_hello_world.py
@@ -16,8 +16,8 @@ import os
 
 import requests
 
-from cai.jumbf import json_to_bytes, json_to_cbor_bytes
-from cai.starling import Starling
+from c2pa.jumbf import json_to_bytes, json_to_cbor_bytes
+from c2pa.starling import Starling
 
 
 # single-claim injection
